requirements.txt
langchain==0.2.15
langchain_core==0.2.37
langchain_openai==0.1.23
langchain_pinecone==0.1.3
langgraph==0.2.16
python-dotenv==1.0.1
streamlit==1.38.0
